+++ ios/Classes/FLTBetterPlayerPlugin.m
@@ -106,9 +106,9 @@ AVPictureInPictureController *_pipController;
 
 - (void)addObservers:(AVPlayerItem*)item {
     if (!self._observersAdded){
+        [_player addObserver:self forKeyPath:@"rate" options:0 context:nil];
         [item addObserver:self forKeyPath:@"loadedTimeRanges" options:0 context:timeRangeContext];
         [item addObserver:self forKeyPath:@"status" options:0 context:statusContext];
-        [_player addObserver:self forKeyPath:@"rate" options:0 context:nil];
         [item addObserver:self
                forKeyPath:@"playbackLikelyToKeepUp"
                   options:0
@@ -167,6 +167,7 @@ AVPictureInPictureController *_pipController;
 
 - (void) removeObservers{
     if (self._observersAdded){
+        [_player removeObserver:self forKeyPath:@"rate" context:nil];
         [[_player currentItem] removeObserver:self forKeyPath:@"status" context:statusContext];
         [[_player currentItem] removeObserver:self
                                    forKeyPath:@"loadedTimeRanges"