+++ lib/src/configuration/better_player_notification_configuration.dart
@@ -1,3 +1,5 @@
+///Configuration of notification which is displayed once user moves app to
+///background.
 class BetterPlayerNotificationConfiguration {
   ///Is player controls notification enabled
   final bool showNotification;