+++ lib/src/configuration/better_player_video_format.dart
@@ -1,3 +1,4 @@
+///Representation of possible video formats in Better Player.
 enum BetterPlayerVideoFormat {
   dash,
   hls,