+++ lib/src/controls/better_player_overflow_menu_item.dart
@@ -1,6 +1,7 @@
 // Flutter imports:
 import 'package:flutter/material.dart';
 
+///Menu item data used in overflow menu (3 dots).
 class BetterPlayerOverflowMenuItem {
   ///Icon of menu item
   final IconData icon;