+++ lib/src/core/better_player_controller_provider.dart
@@ -3,6 +3,7 @@
 import 'package:better_player/src/core/better_player_controller.dart';
 import 'package:flutter/material.dart';
 
+///Widget which is used to inherit BetterPlayerController through widget tree.
 class BetterPlayerControllerProvider extends InheritedWidget {
   const BetterPlayerControllerProvider({
     Key key,