+++ lib/src/playlist/better_player_playlist_configuration.dart
@@ -1,3 +1,4 @@
+///Additional configuration used in Better Player Playlist player.
 class BetterPlayerPlaylistConfiguration {
   ///How long user should wait for next video
   final Duration nextVideoDelay;