+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.52
+* Fixed unregister listener issue in iOS.
+* Updated documentation.
+* [BREAKING_CHANGE] BetterPlayerState visibility changed to private.
+* Fixed HLS audio tracks playlist selection issue.
+* Added enableProgressBarDrag in BetterPlayerControlsConfiguration.
+* Fixed audio track picking in ExoPlayer (Android).
+* Changed default loadingColor.
+
 ## 0.0.51
 * Fixed lint issues.
 * Fixed subtitles setup issue.